ipfs/requirements.txt
# Core Dependencies
Flask==2.3.2
requests==2.31.0
PyYAML==6.0
python-dotenv==1.0.0

# Networking & Performance
urllib3==1.26.16
gunicorn==20.1.0
gevent==22.10.2
httptools==0.6.0

# File Operations
mmap==3.8.1
filelock==3.12.2

# Hashing & Security
cryptography==41.0.3
pyopenssl==23.2.0

# Concurrency
threadpoolctl==3.1.0
concurrent-log-handler==0.9.24

# CLI & Debugging
click==8.1.3
rich==13.4.2

# Optional (for enhanced performance)
uvloop==0.17.0  # (Linux/Mac only)